+++ internal/processor/maps.go
@@ -21,11 +21,13 @@ func CreateMetricSets(samples []interface{}, config *load.Config, i int) {
 	api := config.APIs[i]
 	// as it stands we know that this always receives map[string]interface{}'s
 	for _, sample := range samples {
-
 		// event limiter
-		if (load.EventCount > load.Args.EventLimit) && load.Args.EventLimit != 0 {
-			load.EventDropCount++
-			if load.EventDropCount == 1 { // don't output the message more then once
+		if (load.FlexStatusCounter.M["EventCount"] > load.Args.EventLimit) && load.Args.EventLimit != 0 {
+			load.FlexStatusCounter.Lock()
+			load.FlexStatusCounter.M["EventDropCount"]++
+			load.FlexStatusCounter.Unlock()
+
+			if load.FlexStatusCounter.M["EventDropCount"] == 1 { // don't output the message more then once
 				logger.Flex("debug",
 					fmt.Errorf("event Limit %d has been reached, please increase if required", load.Args.EventLimit),
 					"", false)
@@ -40,7 +42,7 @@ func CreateMetricSets(samples []interface{}, config *load.Config, i int) {
 		// modify existing sample before final processing
 		createSample := true
 		SkipProcessing := api.SkipProcessing
-		for k, v := range currentSample {
+		for k, v := range currentSample { // k == original key
 			key := k
 			progress := true
 			RunKeyRemover(api.RemoveKeys, &key, &progress, &currentSample)
@@ -50,12 +52,10 @@ func CreateMetricSets(samples []interface{}, config *load.Config, i int) {
 				RunValConversion(&v, api, &key)
 				RunValueParser(&v, api, &key)
 				RunPluckNumbers(&v, api, &key)
-				RunSubParse(api.SubParse, &currentSample, key, v)                    // subParse key pairs (see redis example)
-				RunValueTransformer(&v, api, &key)                                   // Needs to be run before KeyRenamer and KeyReplacer
-				RunKeyRenamer(api.RenameKeys, &key)                                  // use key renamer if key replace hasn't occurred
-				RunKeyRenamer(api.ReplaceKeys, &key)                                 // kept for backwards compatibility with replace_keys
-				StoreLookups(api.StoreLookups, &key, &config.LookupStore, &v)        // store lookups
-				VariableLookups(api.StoreVariables, &key, &config.VariableStore, &v) // store variable
+				RunSubParse(api.SubParse, &currentSample, key, v) // subParse key pairs (see redis example)
+				RunValueTransformer(&v, api, &key)                // Needs to be run before KeyRenamer and KeyReplacer
+				RunKeyRenamer(api.RenameKeys, &key)               // use key renamer if key replace hasn't occurred
+				RunKeyRenamer(api.ReplaceKeys, &key)              // kept for backwards compatibility with replace_keys
 
 				currentSample[key] = v
 				if key != k {
@@ -73,8 +73,10 @@ func CreateMetricSets(samples []interface{}, config *load.Config, i int) {
 		if createSample {
 			RunMathCalculations(&api.Math, &currentSample)
 
-			load.EventCount++
-			load.EventDistribution[eventType]++
+			load.FlexStatusCounter.Lock()
+			load.FlexStatusCounter.M["EventCount"]++
+			load.FlexStatusCounter.M[eventType]++
+			load.FlexStatusCounter.Unlock()
 
 			// add custom attribute(s)
 			// global
@@ -143,6 +145,10 @@ func CreateMetricSets(samples []interface{}, config *load.Config, i int) {
 				if api.Prefix != "" && api.Merge == "" {
 					k = api.Prefix + k
 				}
+
+				StoreLookups(api.StoreLookups, &k, &config.LookupStore, &v)        // store lookups
+				VariableLookups(api.StoreVariables, &k, &config.VariableStore, &v) // store variable
+
 				// key filter could be put here
 				AutoSetMetric(k, v, metricSet, api.MetricParser.Metrics, api.MetricParser.AutoSet)
 			}
@@ -226,7 +232,7 @@ func RunKeyRenamer(renameKeys map[string]string, key *string) {
 // StoreLookups if key is found (using regex), store the values in the lookupStore as the defined lookupStoreKey for later use
 func StoreLookups(storeLookups map[string]string, key *string, lookupStore *map[string][]string, v *interface{}) {
 	for lookupStoreKey, lookupFindKey := range storeLookups {
-		if formatter.KvFinder("regex", *key, lookupFindKey) {
+		if *key == lookupFindKey {
 			if *lookupStore == nil {
 				*lookupStore = map[string][]string{}
 			}